requirements.txt
streamlit
pandas
firebase-admin
pyrebase4
fredapi
textblob
setuptools
yfinance
curl_cffi
certifi
requests-cache
matplotlib
investpy 